fix/XMN/pom.xml-version-docker-compose-different-version-Dockerfiles
