+++ docker/docker-compose.override.yml
@@ -1,5 +1,5 @@
-version: "3.9"
-services:
-  postgres:
-    ports:
-      - "5432:5432"
+#name: molly-todolist
+#services:
+#  postgres:
+#    ports:
+#      - "5433:5432"
