+++ docker/flyway/sql/V02__insert_todoItem_data.sql
@@ -1,0 +1,3 @@
+INSERT INTO todo_item VALUES (1, 'David','i am  the first descript','2017-12-28T03:05:56.801Z','2017-12-30T08:05:56.801Z',false);
+INSERT INTO todo_item VALUES (2, 'Oppo','Thinking Machines','2017-12-28T03:05:56.801Z','2017-12-28T03:05:56.801Z',false);
+INSERT INTO todo_item VALUES (3, 'Mac',null,'2017-12-28T03:05:56.801Z','2017-12-28T03:05:56.801Z',false);
