--- docker/sample.env
@@ -1,21 +1,0 @@
-# Project Name
-COMPOSE_PROJECT_NAME=todolist
-
-# Database env
-POSTGRES_DB=postgres
-POSTGRES_USER=${POSTGRES_DB}
-POSTGRES_PASSWORD=${POSTGRES_DB}
-
-
-API_URL:https://api.takicert.takima.io
-
-# Containers url : PUT CORRESPONDING CONTAINERS NAME INSTEAD OF 127.0.0.1 HERE !!
-REDIS_HOST=redis
-BACKEND_URL=http://api:8080
-DB_URL=jdbc:postgresql://postgres:5432/${POSTGRES_DB}
-
-# Flyway
-FLYWAY_USER= ${POSTGRES_USER}
-FLYWAY_PASSWORD= ${POSTGRES_PASSWORD}
-
-
