--- flyway/sql/V2_0__todoItem_data.sql
@@ -1,3 +1,0 @@
-INSERT INTO public.todo_item VALUES (1, 'David','i am  the first descript','2017-12-28T03:05:56.801Z','2017-12-30T08:05:56.801Z',false);
-INSERT INTO public.todo_item VALUES (2, 'Oppo','Thinking Machines','2017-12-28T03:05:56.801Z','2017-12-28T03:05:56.801Z',false);
-INSERT INTO public.todo_item VALUES (3, 'Mac',null,'2017-12-28T03:05:56.801Z','2017-12-28T03:05:56.801Z',false);
