+++ pom.xml
@@ -3,18 +3,18 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
   <parent>
-    <groupId>org.springframework.boot</groupId>
-    <artifactId>spring-boot-starter-parent</artifactId>
-    <version>3.0.3</version>
+      <groupId>org.springframework.boot</groupId>
+      <artifactId>spring-boot-starter-parent</artifactId>
+      <version>3.4.1</version>
     <relativePath/> <!-- lookup parent from repository -->
   </parent>
-  <groupId>io.takima</groupId>
+  <groupId>io.meini</groupId>
   <artifactId>todolist</artifactId>
   <version>0.0.1-SNAPSHOT</version>
   <name>todolist</name>
   <description>To Do List</description>
   <properties>
-    <java.version>19</java.version>
+    <java.version>21</java.version>
     <datasource-proxy.version>1.8.1</datasource-proxy.version>
     <assertj-core.version>3.24.2</assertj-core.version>
   </properties>
@@ -69,7 +69,7 @@
     <dependency>
       <groupId>org.springframework.boot</groupId>
       <artifactId>spring-boot-starter-validation</artifactId>
-      <version>3.0.4</version>
+      <version>3.4.1</version>
     </dependency>
     <dependency>
       <groupId>net.ttddyy</groupId>
@@ -90,7 +90,7 @@
     <dependency>
       <groupId>org.flywaydb</groupId>
       <artifactId>flyway-core</artifactId>
-      <version>9.16.1</version>
+      <version>11.2.0</version>
       <scope>test</scope>
     </dependency>
   </dependencies>
