+++ src/main/java/io/meini/todolist/ToDoListApp.java
@@ -1,0 +1,15 @@
+package io.meini.todolist;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+/**
+ * Hello world!
+ */
+@SpringBootApplication
+public class ToDoListApp {
+    public static void main(String[] args) {
+
+        SpringApplication.run(ToDoListApp.class, args);
+    }
+}
