+++ src/main/java/io/meini/todolist/common/exception/ApiExceptionHandler.java
@@ -1,0 +1,17 @@
+package io.meini.todolist.common.exception;
+
+import io.meini.todolist.common.dto.ExceptionEntity;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+import org.springframework.web.context.request.WebRequest;
+@RestControllerAdvice
+public class ApiExceptionHandler {
+    @ExceptionHandler({ApiException.class})
+    public ResponseEntity<ExceptionEntity> handleApiException(ApiException e, WebRequest request) {
+        e.printStackTrace();
+        return e.getExceptionEntity().toResponseEntity(request);
+    }
+
+}
