--- src/main/java/io/takima/todolist/ToDoListApp.java
@@ -1,15 +1,0 @@
-package io.takima.todolist;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-/**
- * Hello world!
- */
-@SpringBootApplication
-public class ToDoListApp {
-    public static void main(String[] args) {
-
-        SpringApplication.run(ToDoListApp.class, args);
-    }
-}
